+++ include/csutil/scf.h
@@ -230,7 +230,7 @@ void Class::DecRef ()							\
 void Class::scfRemoveRefOwners ()					\
 {									\
   if (!scfWeakRefOwners) return;					\
-  for (int i = 0 ; i < scfWeakRefOwners->Length () ; i++)		\
+  for (size_t i = 0 ; i < scfWeakRefOwners->Length () ; i++)		\
   {									\
     iBase** p = (*scfWeakRefOwners)[i];					\
     *p = 0;								\
